move ExpectedTableOwner to owner package
